# bench: caching bcrypt function for auth