Fixes #1622 Fixes #1640

```
BenchmarkBasicAuth_CacheEnabled-4           3000            369157 ns/op           72023 B/op        451 allocs/op
BenchmarkBasicAuth_CacheDisabled-4            20          66711644 ns/op          509487 B/op       9704 allocs/op
```

Current Master:

```
BenchmarkBasicAuth-4          10         135002061 ns/op         1035518 B/op      19855 allocs/op
```