@dencoded 2018-05-24
left some comments, also it needs to be rebased on master