@asoorm 2018-06-26
```
benchmark                old ns/op     new ns/op     delta
BenchmarkBasicAuth-4     134896502     70577659      -47.68%

benchmark                old allocs     new allocs     delta
BenchmarkBasicAuth-4     19842          10522          -46.97%

benchmark                old bytes     new bytes     delta
BenchmarkBasicAuth-4     1033905       594978        -42.45%
```